# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Haryana Ki Shan Gaming Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  success: boolean;
  message: string;
}

export interface RegisterRequest {
  name: string;
  phone: string;
  password: string;
  referralCode?: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  phone: string;
  upiId?: string;
  walletBalance: string;
  referralCode: string;
  isBlocked: boolean;
  createdAt: string;
}

export interface AuthResponse {
  success: boolean;
  token: string;
  user: UserProfile;
}

export interface MarketItem {
  id: number;
  name: string;
  resultTime: string;
  isActive: boolean;
  latestResult?: string;
  isLive: boolean;
}

export interface MarketsResponse {
  markets: MarketItem[];
}

export interface PlaceBetRequest {
  marketId: number;
  gameType: string;
  number: string;
  amount: number;
}

export interface BetItem {
  id: number;
  marketName: string;
  gameType: string;
  number: string;
  amount: string;
  winAmount?: string;
  status: string;
  createdAt: string;
}

export interface BetResponse {
  success: boolean;
  bet: BetItem;
  newBalance: string;
}

export interface BetsListResponse {
  bets: BetItem[];
}

export interface WalletBalanceResponse {
  balance: string;
}

export interface TransactionItem {
  id: number;
  type: string;
  amount: string;
  status: string;
  description?: string;
  createdAt: string;
}

export interface TransactionsResponse {
  transactions: TransactionItem[];
}

export interface ActiveUpiResponse {
  upiId: string;
  holderName?: string;
}

export interface DepositRequest {
  amount: number;
  upiId: string;
  screenshotUrl?: string;
}

export interface WithdrawRequest {
  amount: number;
  upiId: string;
}

export interface ResultItem {
  id: number;
  marketId: number;
  marketName: string;
  resultNumber: string;
  gameDate: string;
  createdAt: string;
}

export interface ResultsResponse {
  results: ResultItem[];
}

export interface ReferralInfoResponse {
  referralCode: string;
  totalReferrals: number;
  totalEarnings: string;
}

export interface AdminLoginRequest {
  email: string;
  password: string;
}

export type AdminAuthResponseAdmin = {
  id: number;
  name: string;
  email: string;
};

export interface AdminAuthResponse {
  success: boolean;
  token: string;
  admin: AdminAuthResponseAdmin;
}

export interface DashboardStatsResponse {
  totalUsers: number;
  activeUsers: number;
  totalDeposits: string;
  totalWithdrawals: string;
  totalBets: string;
  totalWinnings: string;
  netProfit: string;
  taxAmount: string;
  pendingDeposits: number;
  pendingWithdrawals: number;
}

export interface AdminUsersResponse {
  users: UserProfile[];
}

export interface BlockUserRequest {
  isBlocked: boolean;
}

export interface EditBalanceRequest {
  amount: number;
  operation: string;
}

export interface CreateMarketRequest {
  name: string;
  resultTime: string;
  isActive?: boolean;
}

export interface DeclareResultRequest {
  marketId: number;
  resultNumber: string;
}

export type AdminDepositsResponseDepositsItem = {
  id: number;
  userId: number;
  userName?: string;
  userPhone?: string;
  amount: string;
  upiId: string;
  screenshotUrl?: string;
  status: string;
  createdAt: string;
};

export interface AdminDepositsResponse {
  deposits: AdminDepositsResponseDepositsItem[];
}

export type AdminWithdrawalsResponseWithdrawalsItem = {
  id: number;
  userId: number;
  userName?: string;
  userPhone?: string;
  amount: string;
  upiId: string;
  status: string;
  createdAt: string;
};

export interface AdminWithdrawalsResponse {
  withdrawals: AdminWithdrawalsResponseWithdrawalsItem[];
}

export interface UpiAccountItem {
  id: number;
  upiId: string;
  holderName?: string;
  isActive: boolean;
}

export interface UpiAccountsResponse {
  upiAccounts: UpiAccountItem[];
}

export interface AddUpiRequest {
  upiId: string;
  holderName?: string;
}

export interface BetAnalyticsItem {
  number: string;
  totalAmount: string;
  usersCount: number;
}

export interface BetAnalyticsResponse {
  analytics: BetAnalyticsItem[];
  totalCollected: string;
  totalPayout: string;
  netProfit: string;
}

export type GetBetAnalyticsParams = {
  marketId?: number;
};