# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Haryana Ki Shan Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: requestDeposit
      tags: [wallet]
      summary: Request deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wallet/withdraw:
    post:
      operationId: requestWithdraw
      tags: [wallet]
      summary: Request withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdraw request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
        upiId:
          type: string
        screenshotUrl:
          type: string
      required: [amount, upiId]
    WithdrawRequest:
      type: object
      properties:
        amount:
          type: number
        upiId:
          type: string
      required: [amount, upiId]
...